refactor: 이름 목록 검증 로직 위치 이동
